policy: Re-label metric sources using pkg/source
Previously, this code was relying on the metrics package to provide a
string for prometheus to highlight the source of particular policies.
However, we already have a pkg/source which serves a similar purpose.
Reuse pkg/source for this variable type as well, and extend the variable
type to have a new source type reflecting resources added on the local
API exposed on the node.

Signed-off-by: Joe Stringer <joe@cilium.io>